refactor: create index for utilities
